[update] Set default additional info on confirm email
